This saves a lot of syscalls (20x speedup) at a cost of storing mapping in memory (about 4MB per process) and slightly longer startup.

As measured on my underpowered CPU:

Before:
```
>>> %time from tztrout import tz_ids_for_address
CPU times: user 807 ms, sys: 46.9 ms, total: 854 ms
Wall time: 849 ms

>>> %timeit tz_ids_for_address(country='US', city='New York', state='NY')
1000 loops, best of 3: 403 µs per loop
```

After:
```
>>> %time from tztrout import tz_ids_for_address
CPU times: user 1.32 s, sys: 86.2 ms, total: 1.4 s
Wall time: 1.4 s

>>> %timeit tz_ids_for_address(country='US', city='New York', state='NY')
10000 loops, best of 3: 18.9 µs per loop
```