@wojcikstefan 2018-03-29
> about 4MB per process

How did you arrive at that number? I see ~2-3x of that using some quick profiling tools (granted, I don't fully understand their internals):
```
In [6]: import resource

In [7]: resource.getrusage(resource.RUSAGE_SELF).ru_maxrss / 1024. / 1024.
Out[7]: 111.921875

In [8]: db = InMemoryZipData()

In [9]: resource.getrusage(resource.RUSAGE_SELF).ru_maxrss / 1024. / 1024.
Out[9]: 124.12109375  # ~12MB

In [10]: from pympler import asizeof

In [11]: asizeof.asizeof(db) / 1024. / 1024.
Out[11]: 8.978172302246094  # ~9MB
```