@wojcikstefan 2018-03-29
@tsx please add some before/after perf stats to the PR's description.