@tsx 2018-03-29
> How did you arrive at that number? I see ~2-3x of that using some quick profiling tools.

I did a `sum` on `sys.getsizeof()` on all the deduplicated strings in the db. However I feel even 8-12MB of RAM is a fair trade as opposed to doing quite a few syscalls (open/query/close) on each lookup.

Will test and add more numbers shortly.